api server router  - 단위별 정리
