+++ service/routes/api.js
@@ -2,9 +2,9 @@ const express = require('express');
 
 const router = express.Router();
 
-const loginRouter = require('./login');
-const registRouter = require('./regist');
-const dashboardRouter = require('./dashboard');
+const loginRouter = require('./login/login');
+const registRouter = require('./regist/regist');
+const dashboardRouter = require('./dashboard/dashboard');
 const payRouter = require('./payment/pay');
 
 router.use('/login', loginRouter);
