+++ service/routes/dashboard/creator/creator.js
@@ -1,18 +1,19 @@
 // import
 const express = require('express');
-const pool = require('../model/connectionPool');
-const doQuery = require('../model/doQuery');
-const preprocessing = require('../middlewares/preprocessingData/');
+const doQuery = require('../../../model/doQuery');
+const CustomDate = require('../../../middlewares/customDate');
 
 const router = express.Router();
-const { sortRows } = preprocessing;
-const listOfWithdrawal = preprocessing.withdrawalList;
-const { preprocessingBannerData } = preprocessing;
-const CustomDate = require('../middlewares/customDate');
 
+// sub router
+const bannerRouter = require('./sub/banner');
+const withdrawalRouter = require('./sub/withdrawal');
 
-// 크리에이터 수익금 라우터 및 정보조회
-router.get('/income', (req, res, next) => {
+router.use('/banner', bannerRouter);
+router.use('/withdrawal', withdrawalRouter);
+
+// 크리에이터 수익금 정보조회
+router.get('/income', (req, res) => {
   const { creatorId } = req._passport.session.user;
   const dataQuery = `
   SELECT 
@@ -36,106 +37,8 @@ router.get('/income', (req, res, next) => {
     });
 });
 
-// 크리에이터 광고 내역 라우터
-router.get('/matchedBanner', (req, res, next) => {
-  const { creatorId } = req._passport.session.user;
-  const bannerQuery = `
-  SELECT bm.contractionTime, mi.marketerName, bm.contractionState, br.bannerSrc, bm.contractionId
-  FROM bannerMatched as bm
-  JOIN bannerRegistered as br 
-  ON SUBSTRING_INDEX(bm.contractionId, '/', 1) = br.bannerId
-  JOIN marketerInfo as mi
-  ON SUBSTRING_INDEX(br.bannerId, '_', 1) = mi.marketerId
-  WHERE contractionId LIKE CONCAT('%', ?, '%')
-  ORDER BY contractionTime DESC
-  `;
-  doQuery(bannerQuery, [creatorId])
-    .then((row) => {
-      if (row.result.length > 0) {
-        const result = preprocessingBannerData(row.result);
-        res.send(result);
-      } else {
-        res.end();
-      }
-    })
-    .catch((errorData) => {
-      console.log(errorData);
-      res.end();
-    });
-});
-
-// 크리에이터 현재 광고 중 배너
-router.get('/currentBanner', (req, res, next) => {
-  const { creatorId } = req._passport.session.user;
-  // DB연결후 query문을 통한 데이터 삽입
-  const queryState = `
-  SELECT mi.marketerName, br.bannerSrc
-  FROM bannerMatched as bm
-
-  JOIN bannerRegistered as br
-  ON SUBSTRING_INDEX(bm.contractionId, '/', 1) = br.bannerId
-
-  JOIN marketerInfo as mi
-  ON SUBSTRING_INDEX(bm.contractionId, '_', 1) = mi.marketerId
-
-  JOIN contractionTimestamp as ct
-  ON  ct.contractionId = bm.contractionId
-  
-  WHERE bm.contractionState = 0
-  AND ct.date >= NOW() - INTERVAL 10 MINUTE
-  AND bm.contractionId LIKE CONCAT('%', ?, '%')
-  ORDER BY ct.date DESC
-  LIMIT 1`;
-
-  doQuery(queryState, [creatorId])
-    .then((row) => {
-      const result = row.result.map((value) => {
-        value = Object.values(value);
-        return value;
-      });
-      res.send(result);
-    })
-    .catch((errorData) => {
-      console.log(errorData);
-      res.end();
-    });
-});
-
-router.post('/banner/desc', (req, res) => {
-  const { contractionId } = req.body;
-  const bannerId = contractionId.split('/')[0];
-  const descQuery = `
-  SELECT *
-  FROM bannerRegistered
-  WHERE bannerId = ?`;
-  doQuery(descQuery, [bannerId])
-    .then((row) => {
-      res.send(row.result[0]);
-    })
-    .catch((errorData) => {
-      console.log(errorData);
-      res.end();
-    });
-});
-
-// doQuery 완료
-router.post('/banner/delete', (req, res, next) => {
-  const { contractionId } = req.body;
-  const bannerQuery = `
-  DELETE FROM bannerMatched
-  WHERE contractionId = ? `;
-  doQuery(bannerQuery, [contractionId])
-    .then(() => {
-      res.send([true, '배너가 성공적으로 삭제되었습니다.']);
-    })
-    .catch((errorData) => {
-      console.log(errorData);
-      res.send([false, '배너 삭제에 실패하였습니다 잠시후 시도해주세요.']);
-    });
-});
-
 // 배너 오버레이 URL 주소 가져오기
-router.get('/overlayUrl', (req, res, next) => {
+router.get('/overlayUrl', (req, res) => {
   const { creatorId } = req._passport.session.user;
   const urlQuery = `
   SELECT advertiseUrl, creatorContractionAgreement
@@ -151,9 +54,8 @@ router.get('/overlayUrl', (req, res, next) => {
     });
 });
 
-
 // 수익관리 탭의 크리에이터 별 수익금 차트 데이터
-router.get('/chartdata', (req, res, next) => {
+router.get('/chartdata', (req, res) => {
   // creatorId 가져오기
   const { creatorId } = req._passport.session.user;
   const { dateRange } = req.query;
@@ -268,7 +170,7 @@ router.get('/chartdata', (req, res, next) => {
 });
 
 // creator contraction Update
-router.post('/contraction', (req, res, next) => {
+router.post('/contraction', (req, res) => {
   const { creatorId } = req._passport.session.user;
   const dateCode = new CustomDate().getCode();
   const insertQuery = `INSERT INTO bannerMatched (contractionId)
@@ -292,41 +194,7 @@ router.post('/contraction', (req, res, next) => {
     });
 });
 
-// 크리에이터 출금신청 / 출금신청 금액만큼 creatorIncome에서 제외
-router.post('/withdrawal', (req, res, next) => {
-  const { creatorId } = req._passport.session.user;
-  const withdrawlAmount = req.body.withdrawalAmount;
-
-  const creatorWithdrawalQuery = `
-  INSERT INTO creatorWithdrawal
-  (creatorId, creatorWithdrawalAmount, withdrawalState)
-  VALUES (?, ?, ?)`;
-
-  const creatorIncomeQuery = `
-  INSERT INTO creatorIncome 
-  (creatorId, creatorTotalIncome, creatorReceivable)
-  SELECT creatorId, creatorTotalIncome, creatorReceivable - ?
-  FROM creatorIncome
-  WHERE creatorId = ?
-  ORDER BY date DESC
-  LIMIT 1`;
-
-  Promise.all([
-    doQuery(creatorWithdrawalQuery, [creatorId, withdrawlAmount, 0]),
-    doQuery(creatorIncomeQuery, [withdrawlAmount, creatorId])
-  ])
-    .then(() => {
-      res.send({
-        error: null
-      });
-    })
-    .catch(() => {
-      res.send({
-        error: true
-      });
-    });
-});
-
+// 유저 정보
 router.get('/profile', (req, res) => {
   const profileQuery = `
   SELECT creatorId, creatorName, creatorIp, creatorMail, creatorAccountNumber, creatorContractionAgreement
@@ -355,35 +223,8 @@ router.get('/profile', (req, res) => {
   }
 });
 
-// 크리에이터 출금 내역 불러오기
-router.get('/listOfWithdrawal', (req, res, next) => {
-  // creatorID 가져오기
-  const { creatorId } = req._passport.session.user;
-
-  const listQuery = `
-  SELECT
-  date, creatorWithdrawalAmount, withdrawalState
-  FROM creatorWithdrawal
-  WHERE creatorId= ?
-  ORDER BY date DESC
-  `;
-
-  doQuery(listQuery, creatorId)
-    .then((row) => {
-      if (row.result.length > 0) {
-        const result = listOfWithdrawal(row.result);
-        res.send(result);
-      } else {
-        res.end();
-      }
-    })
-    .catch((errorData) => {
-      console.log(errorData);
-      res.end();
-    });
-});
-
-router.post('/ipchange', (req, res, next) => {
+// ip change
+router.post('/ipchange', (req, res) => {
   const newIp = req.body.value;
   const { creatorId } = req._passport.session.user;
   console.log(req);
@@ -398,7 +239,7 @@ router.post('/ipchange', (req, res, next) => {
     });
 });
 
-// router.post('/welcome', function(req, res, next) {
+// router.post('/welcome', function(req, res ) {
 //   const creatorId = req._passport.session.user.creatorId;
 //   const dateCode =  new CustomDate().getCode();
